+++ src/Maker/skeleton/Entity.tpl.php
@@ -27,39 +27,19 @@ class <?= $class_name ?> implements ResourceInterface, SearchableInterface, Sort
 <?php endif ?>
     private ?int $id = null;
 
-    #[ORM\Column(length: 255)]
-    #[Assert\NotBlank]
-    #[Assert\Length(max: 255)]
-<?php if ($api): ?>
-    #[Groups([self::GROUP_LIST, self::GROUP_READ, self::GROUP_EDIT])]
-<?php endif ?>
-    private ?string $name = null;
-
     public function __toString(): string
     {
-        return (string) $this->name;
+        return (string) $this->id;
     }
 
     public static function getSearchableFields(): array
     {
         // TODO: list the fields matched by `filters[search]` (`relation.field` supported).
-        return ['name'];
+        return [];
     }
 
     public function getId(): ?int
     {
         return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(?string $name): static
-    {
-        $this->name = $name;
-
-        return $this;
     }
 }