+++ src/Maker/skeleton/twig/_form.tpl.php
@@ -2,7 +2,7 @@
 <div class="card shadow-sm mb-4">
     <div class="card-body">
         <div class="row g-3">
-            {{ form_row(form.name) }}
+            {{ form_widget(form) }}
         </div>
     </div>
 </div>