+++ src/Maker/skeleton/twig/index.tpl.php
@@ -2,11 +2,11 @@
 
 {# Column labels are translation keys, looked up in the resource domain then GingermindsCore. #}
 {% set columns = [
-    {label: '<?= $resource->snake ?>.field.name', property: 'name', sortable: true},
+    {label: '<?= $resource->snake ?>.field.id', property: 'id', sortable: true},
     {label: 'common.actions', align: 'end'},
 ] %}
 
 {% block table_row %}
-    <td>{{ item.name }}</td>
+    <td>{{ item.id }}</td>
     <td class="text-end">{{ block('row_actions') }}</td>
 {% endblock %}